requirements.txt
langchain
openai
python-dotenv
webvtt-py
flask
tiktoken
faiss-cpu
flask-cors
langchain-community